# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GUM Internships API
  version: 1.0.0
  description: >
    MVP API contract. Every response uses the envelope {success, data, error, meta}.
    DB is snake_case; this API is camelCase. Auth = Bearer JWT (access 15m, refresh 30d).
    Pagination on list endpoints: ?page=&limit= (default 20, max 100).
    Endpoints marked "Public" need no token. x-required-roles lists allowed roles.
servers:
  - url: https://api.gum-internships.example.com/v1
    description: Production
  - url: https://staging-api.gum-internships.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /catalog/internships:
    get:
      tags: [Catalog]
      summary: 'Public: browse published internships'
      security: []
      parameters:
        - { name: category, in: query, schema: { type: string }, description: Category slug }
        - { name: pricingType, in: query, schema: { type: string, enum: [free, paid, stipend] } }
        - { name: deliveryMode, in: query, schema: { type: string, enum: [recorded, live, hybrid, project_only] } }
        - { name: language, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string }, description: Free-text title search }
        - { name: sort, in: query, schema: { type: string, enum: [newest, price_asc, price_desc], default: newest } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/InternshipSummary' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
  /catalog/internships/{slug}:
    get:
      tags: [Catalog]
      summary: 'Public: internship detail (outline only; preview lessons playable)'
      security: []
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/InternshipDetail' } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /catalog/banners:
    get:
      tags: [Catalog]
      summary: 'Public: live banners by placement'
      security: []
      parameters:
        - { name: placement, in: query, schema: { type: string, enum: [home_hero, home_strip, category_page] } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Banner' } }
  /catalog/pages/{slug}:
    get:
      tags: [Catalog]
      summary: 'Public: published CMS page'
      security: []
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/CmsPage' } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ============================================================ INTERNSHIP AUTHORING
...
  /enrollments/{enrollmentId}/certificate:
    post:
      tags: [Certificates]
      summary: Claim certificate (evaluates certificate_rules; 409 NOT_ELIGIBLE with failed checks)
      x-required-roles: [student]
      parameters:
        - { name: enrollmentId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '201':
          description: Issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Certificate' } }
        '409': { $ref: '#/components/responses/Conflict' }

  # ============================================================ LEARNING
...
  /certificates/{certificateId}/download:
    get:
      tags: [Certificates]
      summary: Signed URL for certificate PDF (owner only)
      x-required-roles: [student]
      parameters:
        - { name: certificateId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { $ref: '#/components/responses/UploadUrlResponse' }
  /verify/{certificateNo}:
    get:
      tags: [Certificates]
      summary: 'Public: verify a certificate by number (rate-limited)'
      security: []
      parameters:
        - { name: certificateNo, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Verification result (valid=false for unknown numbers, still 200)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/CertificateVerification' } }

  # ============================================================ NOTIFICATIONS
...
  /instructor/earnings:
    get:
      tags: [Instructor Finance]
      summary: Earnings ledger (own)
      x-required-roles: [instructor]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, available, settled, reversed] } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Earning' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
...
  /instructor/payouts:
    get:
      tags: [Instructor Finance]
      summary: Settlement history (own)
      x-required-roles: [instructor]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Settlement' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
...
  /admin/settlements:
    get:
      tags: [Admin]
      summary: All settlements
      x-required-roles: [finance_admin, super_admin]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [initiated, processing, completed, failed] } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Settlement' } }
                      meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      tags: [Admin]
      summary: Create settlement run for an instructor's available earnings in a period
      x-required-roles: [finance_admin, super_admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instructorProfileId, periodStart, periodEnd]
              properties:
                instructorProfileId: { type: integer, format: int64 }
                periodStart: { type: string, format: date }
                periodEnd: { type: string, format: date }
                tdsAmount: { type: number, default: 0 }
                notes: { type: string }
      responses:
        '201':
          description: Created with linked earnings frozen
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Settlement' } }
  /admin/settlements/{settlementId}:
    patch:
      tags: [Admin]
      summary: Mark processing/completed/failed; record UTR
      x-required-roles: [finance_admin, super_admin]
      parameters:
        - { name: settlementId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [processing, completed, failed] }
                utrNumber: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
...
  /admin/certificates/{certificateId}/revoke:
    post:
      tags: [Admin]
      summary: Revoke certificate with reason (audited)
      x-required-roles: [moderator, super_admin]
      parameters:
        - { name: certificateId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OkMessage' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    OkMessage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - properties:
                  data:
                    type: object
                    properties: { message: { type: string } }
...
    UploadUrlResponse:
      description: Presigned upload target
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - properties:
                  data:
                    type: object
                    properties:
                      uploadUrl: { type: string }
                      publicUrl: { type: string }
                      headers: { type: object, additionalProperties: { type: string } }
                      expiresAt: { type: string, format: date-time }
...
    NotFound:
      description: 'Resource not found (error.code=NOT_FOUND)'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: 'State conflict, e.g. ALREADY_ENROLLED, BATCH_FULL, SUBMISSION_PENDING_REVIEW'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    Envelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: true }
        data: {}
        error: { type: 'null' }
        meta: { type: object }
    ErrorEnvelope:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, const: false }
        data: { type: 'null' }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: 'Stable machine code, e.g. VALIDATION_ERROR, FORBIDDEN, BATCH_FULL' }
            message: { type: string }
            details: { description: Optional zod issue list }
    PaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    Category:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        slug: { type: string }
        iconUrl: { type: [string, 'null'] }
        displayOrder: { type: integer }
    InternshipSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        slug: { type: string }
        shortDescription: { type: [string, 'null'] }
        category: { $ref: '#/components/schemas/Category' }
        pricingType: { type: string, enum: [free, paid, stipend] }
        price: { type: number }
        currency: { type: string }
        deliveryMode: { type: string, enum: [recorded, live, hybrid, project_only] }
        paceType: { type: string, enum: [batch, self_paced] }
        durationWeeks: { type: [integer, 'null'] }
        languages: { type: array, items: { type: string } }
        thumbnailUrl: { type: [string, 'null'] }
        instructorName: { type: string }
        publishedAt: { type: [string, 'null'], format: date-time }
    InternshipDetail:
      allOf:
        - $ref: '#/components/schemas/InternshipSummary'
        - type: object
          properties:
            description: { type: [string, 'null'] }
            outcomes: { type: array, items: { type: string } }
            prerequisites: { type: array, items: { type: string } }
            gstRate: { type: number }
            totalWithGst: { type: number, description: Server-computed display price }
            certificateRules: { type: object }
            curriculum:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer, format: int64 }
                  title: { type: string }
                  lessons:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer, format: int64 }
                        title: { type: string }
                        type: { type: string, enum: [video, live, document, quiz] }
                        durationMinutes: { type: [integer, 'null'] }
                        isPreview: { type: boolean }
            upcomingBatches: { type: array, items: { $ref: '#/components/schemas/Batch' } }
            instructor: { type: object, properties: { name: { type: string }, bio: { type: [string, 'null'] }, expertise: { type: array, items: { type: string } } } }
...
    Batch:
      type: object
      properties:
        id: { type: integer, format: int64 }
        internshipId: { type: integer, format: int64 }
        name: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        enrollmentDeadline: { type: [string, 'null'], format: date-time }
        seatsTotal: { type: integer }
        seatsFilled: { type: integer }
        seatsLeft: { type: integer }
        waitlistEnabled: { type: boolean }
        status: { type: string, enum: [scheduled, enrolling, ongoing, completed, cancelled] }
...
    Certificate:
      type: object
      properties:
        id: { type: integer, format: int64 }
        certificateNo: { type: string }
        internshipTitle: { type: string }
        status: { type: string, enum: [issued, revoked] }
        issuedAt: { type: string, format: date-time }
        pdfUrl: { type: [string, 'null'], description: Signed URL when served }
        verifyUrl: { type: string }
    CertificateVerification:
      type: object
      properties:
        valid: { type: boolean }
        certificateNo: { type: string }
        status: { type: string, enum: [issued, revoked] }
        learnerName: { type: [string, 'null'] }
        internshipTitle: { type: [string, 'null'] }
        issuedAt: { type: [string, 'null'], format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
...
    Earning:
      type: object
      properties:
        id: { type: integer, format: int64 }
        internshipTitle: { type: string }
        orderNo: { type: string }
        grossAmount: { type: number }
        revenueSharePercent: { type: number }
        amount: { type: number }
        status: { type: string, enum: [pending, available, settled, reversed] }
        availableAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...
    Settlement:
      type: object
      properties:
        id: { type: integer, format: int64 }
        settlementNo: { type: string }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        grossAmount: { type: number }
        tdsAmount: { type: number }
        payableAmount: { type: number }
        status: { type: string, enum: [initiated, processing, completed, failed] }
        utrNumber: { type: [string, 'null'] }
        paidAt: { type: [string, 'null'], format: date-time }
...
    Banner:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        imageUrl: { type: string }
        linkUrl: { type: [string, 'null'] }
        placement: { type: string, enum: [home_hero, home_strip, category_page] }
        displayOrder: { type: integer }
    CmsPage:
      type: object
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string }
        title: { type: string }
        contentMd: { type: string }
        metaTitle: { type: [string, 'null'] }
        metaDescription: { type: [string, 'null'] }
...